+++ src/app/(auth)/sign-in/page.tsx
@@ -5,7 +5,6 @@ import { protectPage } from "@/features/auth/queries";
 
 export default async function SignInPage() {
   const user = await protectPage();
-
   if (user) redirect("/");
 
   return <SignInCard />;
