+++ src/app/(auth)/sign-up/page.tsx
@@ -5,7 +5,6 @@ import { protectPage } from "@/features/auth/queries";
 
 export default async function SignUpPage() {
   const user = await protectPage();
-
   if (user) redirect("/");
 
   return <SignUpCard />;
