+++ src/features/auth/api/use-current.ts
@@ -2,6 +2,10 @@ import { useQuery } from "@tanstack/react-query";
 
 import { client } from "@/lib/rpc";
 
+/**
+ * @description Hook for getting the current user
+ * @returns Query function
+ */
 export const useCurrent = () => {
   return useQuery({
     queryKey: ["current"],
