+++ src/features/auth/schemas.ts
@@ -10,6 +10,9 @@ const numberErrorMessage = "Password must contain at least one number";
 const specialCharacterErrorMessage =
   "Password must contain at least one special character";
 
+/**
+ * @description Schema for password validation
+ */
 export const passwordSchema = z
   .string()
   .min(8, { message: minLengthErrorMessage })
@@ -25,6 +28,15 @@ export const passwordSchema = z
     message: specialCharacterErrorMessage,
   });
 
+/**
+ *  @description Schema for signing in
+ * @example
+ * ```ts
+ * const data = signInSchema.parse({
+ *  email: "john.doe@example.com",
+ *  password: "password123",
+ * });
+ */
 export const signInSchema = z.object({
   email: z.string().email({ message: "Invalid email address" }),
   password: z
@@ -32,6 +44,17 @@ export const signInSchema = z.object({
     .min(8, { message: "Password must be at least 8 characters" }),
 });
 
+/**
+ * @description Schema for signing up
+ * @example
+ * ```ts
+ * const data = signUpSchema.parse({
+ *  name: "John Doe",
+ *  email: "john.doe@example.com",
+ *  password: "password123",
+ * });
+ * ```
+ */
 export const signUpSchema = z.object({
   name: z.string().trim().min(1, { message: "Name is required" }),
   email: z.string().email({ message: "Invalid email address" }),
