+++ src/features/workspaces/components/create-workspace-modal.tsx
@@ -2,7 +2,7 @@
 
 import { ResponsiveModal } from "@/components/responsive-modal";
 
-import { useCreateWorkspaceModal } from "../hooks/use-creat-workspace-modal";
+import { useCreateWorkspaceModal } from "../hooks/use-create-workspace-modal";
 import { CreateWorkspaceForm } from "./create-workspace-form";
 
 export function CreateWorkspaceModal() {
