+++ src/features/workspaces/schemas.ts
@@ -1,5 +1,15 @@
 import { z } from "zod";
 
+/**
+ * @description Schema for creating a workspace
+ * @example
+ * ```ts
+ * const data = createWorkspaceSchema.parse({
+ *  name: "New Workspace",
+ *  image: new File([], "image.png"),
+ * });
+ * ```
+ */
 export const createWorkspaceSchema = z.object({
   name: z.string().trim().min(1, { message: "Workspace name is required" }),
   image: z.union([
@@ -11,6 +21,16 @@ export const createWorkspaceSchema = z.object({
   ]),
 });
 
+/**
+ * @description Schema for updating a workspace
+ * @example
+ * ```ts
+ * const data = updateWorkspaceSchema.parse({
+ *  name: "New Workspace",
+ *  image: new File([], "image.png"),
+ * });
+ * ```
+ */
 export const updateWorkspaceSchema = z.object({
   name: z
     .string()
@@ -25,6 +45,16 @@ export const updateWorkspaceSchema = z.object({
       .optional(),
   ]),
 });
+
+/**
+ * @description Schema for joining a workspace
+ * @example
+ * ```ts
+ * const data = joinWorkspaceSchema.parse({
+ *  code: "123",
+ * });
+ * ```
+ */
 export const joinWorkspaceSchema = z.object({
   code: z.string().trim().min(1, { message: "Invite code is required" }),
 });
